Make sure clearBackStack clears fragment states
Previously calling clearBackStack on the navController would attempt to
restore and immediately pop the saved state. We were failing to pass
down the restore option and therefore we never restored and popped the
fragment state.

We need to add the correct flags to the clearBackStack calls.

RelNote: "NavController will now correctly clear the back fragment
manager back stack state when calling `clearBackStack()`."
Test: added navControllerWithFragmentTest
Bug: 271190202

Change-Id: Ic1cce69633d6f954f7a2f2600a7426bbb241713f