Remove priv classes from provider interface
This change is due to change in public branch - aosp/2461615

We are now adding an origin setter on the
top level create and get requests in androidx.credentials
so we don't need special priv classes on the provider side too.

This change needs to go in before the prov classes under
credentials.androidx can be removed so that there are no downstream
build issues when merging from androidx-main.

Test: built locally
Relnote: "Deleting priv API classes. Instead developers should use
origin setters to make get/create requests on behalf of other apps."

Change-Id: I17b2cd337736825a8a50ef342a10d23f61e16db8