Fix paragraph.getLineForVertical
getLineForVertical(position) added the top padding
instead of removing it.

This surfaced when includeFontPadding was made
off by default in aosp/2466791.

Material3 TabTest relied on device density/font. Updated
the test to be more correct when padding is false. Did
not add test fonts yet to these tests since it would be
the first and it would change the dependencies of the
project.

Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT

Bug: 228583901
Change-Id: I41ae7ae4aabcbf4f471dced58d4f75fd0bc0035a