Don't use System.currentTimeMillis() in WorkSpec#calculateNextRunTime()
lastEnqueueTime is 0 in some synthetic tests where WorkSpec in
instantiated via constructor.

In prod it can be the case when work chains are scheduled. Work with
unmet dependencies is not considered 'enqueued' until all its
dependencies are met.

In current scenario, if this method were called for that work, then it
would always return a time in the future if it had any delay (so the work never completes).
Preserve that behavior by returning a schedule far in the future.

Test: Existing WorkSpec tests pass.
Change-Id: Icb1565e1e7c3a216371e30d47860fe10a78cba0a