JS evaluation returns data via pipe and not Binder
This CL uses pipes to return data to the sandboxed process instead of
relying on Binder. The return/error values can be large and as a result,
there was a risk of overflowing the Binder transaction buffer leading to
TransactionTooLargeException. Even though the buffer has a maximum size
of 1MB, this is shared with the entire process and the actual limit is
dependent on runtime parameters.

A new flag has been added for clients to check if this new behaviour is
available in the underlying WebView. Currently, the flag has a ":DEV"
suffix. This will be removed once development is complete.

Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroidTest
Bug: 269428261
Change-Id: I03f2dfcb215d53ecd9d22937d866939cf9d9dabc