Fix Visibility.Gone not working in certain cases
Visibility.Gone would only work when using a fixed dimension on a
Composable. For example, when using WrapContent, the layout around the
usage of Gone would be correct, but the Composable would still paint and
retain its WrapContent dimensions.

This addresses that issue by skipping placement of Composables marked as
Gone.

Fixes: 272110322
Test: updated
Change-Id: I4fd3bc63e9a3e2eb01be82e65b6017dd45ee281e