Date pickers API changes for selectable dates
This CL removes the lambda for date validation and introduce a
`SelectableDate` interface that can be set on the state when calling
`rememberDatePickerState` or `rememberDateRangePickerState`.
The `SelectableDate` provides functions for determining if a date or a
year should be selectable and enabled in the UI.

Using this interface, the UI can now display disabled years when
selecting a year in the DatePicker, or when navigating to a month within
a non-selectable year.

Bug: 271912022
Test: DatePickerTest, DateRangePickerTest, screenshot tests.
Relnote: "Removed the dateValidator from DatePicker and DateRangePicker
and introduced a SelectableDates interface that can be set on the state
in order to control which dates or years are selectable in the UI."

Change-Id: Ic2fc6f681417f632e7279016bd67ccbe372d653e