Ensure FragmentNavigator handles back properly
We need to ensure that FragmentNavigator handles all cases of system
back, with and without animators.

Fragments have two different cases when it comes to specialEffects.
Animators and Transitions behave the same, which is Fragments consider
their status before making Lifecycle changes. Animations are handled the
same as no special effects and Fragments ignore them altogether.

Previously, when FragmentNavigator attempted to handle system back, if
the Fragment ignored the special effects, it was already destroyed and
we popped the entry, but if it did not ignore the special effects, we
had no way of knowing the fragment was going away and we failed to pop
the entry.

Now the timing of the callback used by FragmentNavigator has been
changed (see aosp/2486363), we need to make sure both cases are handled
properly.

RelNote: "`FragmentNavigator` will now properly pop entries when the
back stack is popped via system back or `popBackStack()` and whether or
not the transaction uses effects for the fragment."
Test: Added OnBackPressedTest

Change-Id: I81bdf764c4dfb070b74017f00bbd7ba740f831c4