Add change tracking to MutableTextFieldValue.
Introduces the ChangeList interface which exposes a list of
non-overlapping, non-adjacent changes in terms of the range in the
original text and the corresponding range in the changed buffer.

ChangeTracker is an internal class that implements ChangeList and does
all the actual tracking of offsets. Both MutableTextFieldValue and
EditProcessor use ChangeTracker to track changes made to their internal
text buffer, and the ChangeList is exposed via a property on
MutableTextFieldValue.

Test: ChangeListTracker
Test: BasicTextField2Test
Test: TextFieldStateTest
Test: manually tested samples and demos
Relnote: n/a
Change-Id: I8dfda83e59b8f0cbe97cb9f97790a9a201cbb754