Added flavor generation tasks and fixed global one in agp 8.1.0
Due to b/265438201, the global generation task does not trigger
generation for all the variants but only for the ones with build type
`release. So calling `generateBaselineProfile` is equivalent to
`generateReleaseBaselineProfile`. Additionally we cannot create flavor
generation tasks (such as `generateFreeBaselineProfile`) because it'd
build for multiple build types. This PR fixes the global generation task
ensuring that the generation tasks for all the variants for every build
type run when calling `generateBaselineProfile` and adds flavor
generation tasks.

Fixes: 262070052
Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Change-Id: I2c30ff7849dd7ee7353f2665e2992b1496b3521f