Make a view important to a11y when adding an accessibility delegate.
This will apply to IMPORTANT_FOR_ACCESSIBILITY_AUTO views. If a
developer is adding a delegate, they likely  want this view to be
visible to services (and should set importance to NO if not).

Remove the code that checks for an ancestor that hides descendants.
If the ancestor is hidden or hides descendants, the subtree rooted
at the ancestor is guaranteed to not be exposed to accessibility
services. This also aligns more with View#setImportantForAccessibility,
which doesn't check the ancestors.

Relnote: "Made a view important for accessibility if the view has an
accessibility delegate."

Test: AccessibilityDelegateCompatTest
Bug: 275751628
Change-Id: If2b54e94daf9612e84aab13f498f8028ee8d68d6