Add ItemImageShape property to GridTemplate.
The ItemImageShape will be used by the host to crop all grid item images
into the specified shape. For now only UNSET (don't
crop the image) and CIRCLE are supported.

This is an experimental feature that is part of the Car App Api level 7.

Bug: 265202765
Test: unit tests
Relnote: "Add ItemImageShape property to GridTemplate"
Change-Id: Ibf431e8965f506395e376d1fc2f7cfd10caaa902