Get TextFieldValue out of public BasicTextField2 API.
Replace it with a new type, called TextFieldCharSequence. This is almost
identical to TextFieldState with one crucial difference: instead of
having a text property, it is itself the text. TextEditState implements
CharSequence. This removes a level of indirection required to get the
actual text out of a TextEditState: `state.value.text` is now just
`state.value`.

Furthermore, because it doesn't expose a final, immutable
String/AnnotatedString, it gives us the freedom to change how the text
is actually stored in the future, opening the possibility of re-using
internal data structures to avoid copying.

Test: TextFieldCharSequenceTest
Test: existing tests
Relnote: n/a
Change-Id: Ic8d736ff805e620038c011b7cf8700eeab96c26e