Roll forward: TestImageUtil#calculateColorDiff does not calculate the diff
The bitwise and substraction is executed in a wrong order.

- Fixed the calculation.
- Move JpegBytes2CroppedBitmapTest to instrument test. Unit test cannot verify the content of the image.

The original CL(aosp/2520635) failed a test failure. It looks like on that specific emulator, if the image is encoded to JPEG and decoded back to Bitmap, the color will be slightly different. The solution is to skip emulator/API 21 since they are known to be unstable.

Bug: 270190514, 276967642
Test: manual test and ./gradlew bOS
Change-Id: I6a3eb143c01f86a353e67d2777d39fea82b568fe