Confine lookahead pass to only the content in LookaheadScope
LookaheadScope previously initiated a lookahead pass from the
parent LayoutNode. While this allows lookahead placement of
the immediate child nodes in the scope, it also comes with
2 caveats: 1) Siblings of LookaheadScope will also get
a lookahead pass, which may come as a surprise. 2) Also
the deal breaker: SubcomposeLayout may have a LookaheadScope
containing some but not all of the children. The subcomposition
would be initiated in a measure pass. When the lookaheadScope
gets subcomposed, the SubcomposeLayout would need a lookahead
pass, which breaks the contract that lookhead pass always
comes before measure pass.

This change avoids propagating lookahead pass to the parent
of the LookaheadScope for the two reasons mentioned above.

Test: updated

Change-Id: I946c61161caa27ef7ee1bc1f65c32ab32057cc70