Deprecate LazyListScope#itemsIndexed extension for LazyPagingItems
This method is deprecated because index-based keys for non-null items is susceptible to error when prepending, as prepends would shift item indices.

The ReplaceWith function will direct users to the original LazyListScope.items function and auto populate the replacement with current key and contentType params. This will, however, result in compilation error as the current key/contentType is a lambda that passes in both index and item, while the replacement lambda only passes in item. Users will have to refactor the replaced code to remove use of index.

If neither key nor contentType param is supplied, it will use the default implementation of each one.

Test: compose demos
Test: ./gradlew paging:paging-compose:cC
Bug: 276989796
Relnote: "The LazyListScope.itemsIndexed extension inside LazyPagingItems has been deprecated. Replaced with LazyListScope#items through LazyPagingItems's #itemCount, #itemKey, and #itemContent helpers."
Change-Id: I92c8ffc330fd98f9fb31fe14db2da7a3f7a3b547