[GH] Replace usage of TestDispatcher with StandardTestDispatcher in paging-compose
In Multiplatform Paging, the classes `DirectDispatcher` and `TestDispatcher` in `testutils-ktx` had to be multiplatformized to allow dependent tests in `paging-compose`, `paging-common`, and `paging-runtime` to be moved to the respective `commonTest` folders.

I'm attempting to replace all usages of `DirectDispatcher` and `TestDispatcher` that Multiplatform Paging depends on with those found in `kotlinx.coroutines.test`, so hopefully the `testutils-ktx` dependency no longer has to be multiplatformized in my fork.

I'm starting this effort by doing it with `paging-compose`.

Test: ./gradlew test connectedCheck
Bug: 270612487

This is an imported pull request from https://github.com/androidx/androidx/pull/505.

Resolves #505
Github-Pr-Head-Sha: 002b706d42c6ce25718ccf1b3745872f27eb82d9
GitOrigin-RevId: 2b3227350681a7cc3ae43b1a52b311d488b5752e
Change-Id: Id018cf4bf64dac3154eff2ea98e6145aff4290fc