Add SoftwareKeyboardControllerCompat
Extracts the implementation of WindowInsetsControllerCompat.show and
hide for the IME type into a new SoftwareKeyboardControllerCompat type.
While WindowInsetsControllerCompat needs the correct Window instance for
all of its functionality, the implementations (and workarounds) for
showing and hiding the IME only require a View instance. By extracting
these out, other implementations that want to show and hide the IME
(such as Compose) can only need the View, and not the correct Window,
which can be difficult when dealing with Dialog interoperability.

Test: New
Change-Id: I2740d94a950bad76c961975ec1899f747b203229