Make BeginGetCredentialOption abstract
BeginGetCredentialOption is to be abstract to match the pattern being
followed by BeginCreateCredentialReuest and CredentialOption. Any custom
class can extend from BeginGetCustomOption.
In addition in this change, we are removing restrictions from properties on this base class
and making them public so as to avoid overriding attributes in
BeginGetCustomCredentialOption. Overriding properties in constructions
is discouraged in kotlin guidelines and causes undefined behavior.

Test: local build + unit tests
Bug: 279500807
Relnote: "Make BeginGetCredentialOption abstract"

Change-Id: If1ba2e5ae7f8267ff0ca52ac851433f0d21878fd