Fix preview rotation in core test app
Currently core test app depends on texture rotation and display rotation to correct the preview.  This is no longer true when preview is the output of an internal OpenGL renderer. In that case the texture rotation will be 0 and display rotation might already be applied. In that case, we depend on the rotation provided by SurfaceRequest.

My theory is that there is a bug in the MVP transformation that only happens to work if the texture rotation is 90/270. It works because we apply it in the wrong direction and it cancels out with the 180 rotation in the texCoords. When StreamSharing is enabled, texture rotation becomes 0 and this breaks. The fix is apply it in the right direction and don't rotate in texCoords.

Tested on front/back camera with and without stream sharing.

Bug: 269613528
Test: manual test and ./gradlew bOS
Change-Id: Ic945c57f4f94aefe2d142fefdabfcf9fe3eecd3f