jsengine: Change guava to API dependency
ListenableFuture is part of the javascript engines public API
so we need to declare it as such.
We are currently forcing consumers to redeclare this dependency.

Initially tried just making guavaListenableFuture an API
dependency but it seems like this doesn't work properly
and from some brief search it seems this is now considered a mistake
to break out. For these reasons, sticking with the general
guava android lib as the api dependency.

Test: archived maven local and used this in a separate test app
Bug: 249257418
Change-Id: I99e7ab3b4fd851b3dd5eff982b73c4facbc08fb5