Introduce the SdkActivityLauncher.
This interface allows app developers to control when/how SDKs launch
their activities. The interface can be turned into a bundle for
transport, so it can be used by SDKs in the Privacy Sandbox.

The provided implementation of the interface disposes the activity
reference used to start SDK activities to avoid leaks, since the object
will be retained by Binder threads.

Relnote: "Introduce the SdkActivityLauncher. An interface for allowing
runtime enabled SDKs to launch activities."

Test: ./gradlew :privacysandbox:ui:ui-{client,provider}:test --parallel

Change-Id: I5b3dc443f76c5c43e88a43ba272869d9b15668b0