Support non-JPEG output in the new ImageCapture pipeline.
Migrate non-JPEG output to the new pipeline. This code path is only used by tests. This change should not affect any public users.

- In ImagePipeline and CaptureNode, separate output format from input format.
- Plumb the output format through ProcessingNode. If the output format is non-jpeg, deliver result without processing it.

This also fixed an issue in aosp/2552990. We should return ImageProxy back to app, even if ImageReaderProxyReader is used. The tests was false positive because non-JPEG format was disabled.

Bug: 279046948
Test: manual test and ./gradlew bOS
Change-Id: I5a887dc9b0b3517fbaaf7855501ed245dd9b5e9a