Fix variance resolver when used with asMemberOf.
This CL fixes a bug in `KSTypeVarianceResolver` where we would get
incorrect wildcards if the `XType` called with `#asMemberOf(XType)`
does not match the scope of the type itself. This change requires
passing in the asMemberOf type when creating the scope.

This CL also adds a lot more coverage for `KSTypeVarianceResolver`
when used with types that are resolved from type variables (see
`KSTypeVarianceResolverTest.kt`.

Test: KSTypeVarianceResolverTest.kt
Change-Id: Ica3cea5b0c16ccd4e5616330b98ab29669742fd2