Dispatch onBackStackChange callbacks to the correct records
With reorderingAllowed true, it is possible to get in a scenario where
there are multiple transactions in the FragmentManager, some added to
the back stack and others not. FragmentManager currently looks at the
global result and if a single transaction changes the back stack it
dispatches onBackStackChange callbacks for all of the records.

We need to ensure only the individual records that explicitly add things
to the back stack cause the onBackStackChange callbacks to be
dispatched.

RelNote: "Fixed an issue where the onBackStackChangeListener callbacks
were being dispatched for transactions that did not actually change the
back stack if they were mixed in with transactions that do."

Test: Added onBackStackChangedListenerTest
Bug: 279306628
Change-Id: I0eb5c9622f2984232ac106bb975ba72ae9b2435f