Exposed cached data via PagingData
When using Flow<PagingData<T>>.cachedIn(scope), data that has been collected within that generation is cached for as long as the scope remains alive.
Normally the cachedData can only be accessed by new subscriptions when they collect on the SharedFlow returned by cachedIn. Now we expose this cachedData via PagingData so that the cached data can be accessed synchronously without requiring collection.

Test: ./gradlew paging:paging-common:test
Bug: 177245496

Change-Id: I36a894d7d90f61396604e7847c5be6027b5e0991