Generate API levels JSON for AndroidX
Updated to do filtering of which files to use at configuration time.

This filters API files so only the latest one from each version is used (e.g. if beta01, beta02, and beta03 all exist, only beta03 will be used) and rounds up the name of the API versions to stable for past versions.

Sample generated JSON files for :core:core and :fragment:fragment are in this drive folder (as "core:core-filtered.json" and "fragment:fragment-filtered.json"):

Bug: 281719231
Bug: 281728093
Test: `./gradlew generateApi` runs successfully, validation described above, unit tests added in GenerateApiLevelsTest

Change-Id: I7d0b73bf89c9849f65819abd236e4c672612e5d0