Skip tests in CameraControllerFragmentTest if necessary UiObject can't be found
Issue b/277179455 can't be reproduced locally.

The necessary UiObjects for some tests can not be found when running the tests. It might be caused by the activity is covered by other popup dialog or may be something wrong in UiDevice#findObject(). No matter which is the root cause, the tests are unnecessary to run when the UiObject can't be found. Therefore, skip the test when the UiObject can't be found.

Bug: 277179455
Test: CameraControllerFragmentTest on Nexus 5
Change-Id: I2c5faa6d5d1aec3d25efb31d276ceb35d341ea03