Add PagingSourceFactory interface
This functional interface allows creating a PagingSourceFactory by passing in a lambda that returns a PagingSource, or by creating a class that extends this interface. The factory can be passed into Pager as the pagingSourceFactory.

Test: ./gradlew paging:paging-common:test
Bug: 280655188
Relnote: "Added a new PagingSourceFactory functional interface. Override the invoke() method with an implementation that returns a new instance of PagingSource. This factory can be used to instantiate a Pager."
Change-Id: I3316590a0d667566edeb40d4e99e005cd6cee3ac