Optimize first time vector rendering
A PathComponent would always allocate two Path instances:
- The working path
- A render path

In the common case however, when there's no trimming, both paths are
the same, but a copy is required to match render path and the source
path. This change modifies how the render path is handled to use a
reference instead when there's no trim, avoiding an allocation and
a copy.

This change also doesn't set isTrimDirty to true by default as it
would cause updateRenderPath() to be called twice on the first two
draws (causing 2 copies!). Since the dirty flag is always set when
changing the default trim values, it is not necessary to set it to
true by default.

Lastly this change makes a few cosmetic changes to make property
and function declarations more consistent by ommitting the type when
it is not necessary to be declared.

Test: VectorTest
Test: compose.integration-tests.demo
RelNote: Optimize Vector memory usage and first-frames rendering
Change-Id: I2f3c67bf8ea1cd09c6d782f19a2dae5de73b3889