Refactor KSTypeVarianceResolver
This CL refactors KSTypeVarianceResolver but shouldn't cause any
functional changes to the code.

The main benefit of this change is that it removes a lot of the logic
that was repeated in many of the functions. Now that logic is
encapsulated in the KSTypeWrapper and KSTypeArgumentWrapper classes and
can be shared across all functions.

Another benefit of this change is that the wrappers themselves are now
much simpler to think about because they no longer extend KSType and
KSTypeArg, respectively. This makes easier to reason about where each
method call is coming from since we don't have to think about the
superclass and overriding.

Test: KSTypeNamesGoldenTest.kt
Change-Id: Ie5a108db6b9aadcc66f334b58acb3c36d8d7bf87