Fix test flakiness in JankStatsTest
There has been significant test flakiness in JankStatsTest, particularly
with testNoJank, testFrameStates, and testComplexFrameStateData. This
situation tends to occur on emulators, and especially in some runtime
environments, such as emulators running in the test lab.

The root cause is that the system sometimes does not display (and process
JankStats data for) all frames that are rendered, resulting in a mismatch
of expectations and results, such as the number of janked frames and the
state data on each resulting frame.

The fix is to adjust the expectations, to allow for some minor frame skips.
For example, we no longer require the exact number of jank frame results as
the number of frame the test supposedly created. Also, for
testComplexFrameStateData, the test allows for resulting data to be on either
the current expected frame or the next one, allowing for a single skipped frame.

Bug: 270721984
Bug: 238890043
Test: JankStatsTest tests all pass, locally and on test lab emulators
Change-Id: I034b9c1b6acf6a8c817d0edf7708e4249351941b