Fix interruption in lazy placement animations
Previously I broke this logic in aosp/2464372. We can't call snapTo() before doing the animation when another animation is in progress as it breaks the continuity of the animations, and passing the correct velocity to the new animation is not enough. In cases like when the previous item is changing their size on every frame it caused the animation to not happen at all until the resize animation stop. With new logic we nicely update the animation target on each frame, as it was before the recent refactoring.

Test: new test in all three LazyFooAnimateItemPlacementTest
Bug: 280660538
Change-Id: I428163cda9be1bc333d0d50a2d474dba45762af2