Pager to consume pre drag deltas from children when it is not settled.
To make sure we don't end up in situations where list in a page is only partially visible and scrolls, pager should make sure it is at the `currentPAgeOffsetFraction == 0` before allowing children to proceed with scrolling.

Relnote: default PagerDefaults.pageNestedScrollConnection now makes sure that pager is settled before allowing children to scroll. This prevents situations when pager's child is scrolling, but not fully visible.
Fixes: 211793934
Test: added new
Change-Id: I9dfba5494483ebe0adca07326068611b4a98ae1a