Don't retain snapshot reference in ComposerImpl
Avoid holding a reference to the current snapshot after composition
completes. This breaks a reference chain originating from the
Recomposer's writeObserver registered with the composition snapshot that
can cause some data to still be referenced until the next recomposition
or until the composition is disposed.

This is the first of two related fixes; committed/disposed snapshots
should also drop references to their observers, as those observers may
still hold references to other data making holding a long-lived
reference to a disposed Snapshot a potential source of surprising
behavior.

Fixes: 262356264
Test: JvmCompositionTests.compositionDoesNotRetainSnapshotReference
Change-Id: I108c3be00f497f03fe993f986df7bf900699ebaa