Fix KSTypeVarianceResolver for raw TypeName
In a previous CL, I removed the check for `isJavaRawType()` since I
thought it was already covered by the `needsWildcardResolution()`check:

```
val needsWildcardResolution: Boolean by lazy {
  annotated.isInKotlinCode() || container?.isInKotlinCode() == true
}
```

In particular, if a raw type is used then it means that our scope is not
within kotlin code. However, that assumption was apparently wrong (still
not sure why) so I'm adding back the check for `isJavaRawType()` and
also adding a regression test.

Test: XTypeTest.kt
Change-Id: I0956f730be3f22f6b909b24e4f679f9b92a8fc96