Remove checkAidlApiRelease to match intended AIDL versioning policy
This is specific to Jetpack since we ship to Google apps from tip-of-tree,
but it might not make sense for all clients of the plugin. In the future,
we might need to add an API that allows users to set up a checkApi task
for a specific (e.g. last stable, released version) directory.

Bug: 282781188
Test: ./gradlew :core:core:checkAidlApi
Change-Id: I1916eeef9c76ff1f9abe4862766a5d03fcdc9820