Fix ActivityOptions for LazyList actions
Stores the ActivityOptions bundle for a list item action in the
trampoline intent, and uses it when launching the target activity.

Adds a new API to provide ActivityOptions for LazyColumn and
LazyVerticalGrid that will be used for all actions in the list.

Bug: 283466829
Test: manual test with ScrollableAppWidget demo that target activity is
 launched with the given options bundle, added unit tests for LazyColumn
 and LazyVerticalGrid
Relnote: "Added a new API to provide ActivityOptions for LazyColumn and
 LazyVerticalGrid that will be used for all actions in the list."
Change-Id: Id8d710b4f5c9fa02f4a76311d14b77462fc76ece