Finalize session when we fail to build output configurations
When we fail to build the output configurations, it means the particular
CaptureSessionState has failed to be configured. As such, it should be
finalized accordingly.

Note that this causes issues becaues in this case, CameraDeviceWrapper
would not have received the CaptureSessionState, and thus it would not
be able to finalize the failed CaptureSessionState.

Bug: 285034336
Test: CameraPipe builds
Change-Id: I06587c411ade1a0bd0a2a7b2823d1e0e8dbbc45e