Adding Material3 Stepper to Compose for Wear OS
Stepper allows users to make a selection from a range of values.
It's a full-screen control with increase button on the top, decrease
button on the bottom and a slot in the middle.
Value can be increased and decreased by clicking on the increase and
decrease buttons.
This version doesn't have range semantics by default. It can be added
using modifier such as Modifier.rangeSemantics.

Test: androidx.wear.compose.material3.Stepper tests
Bug: 266370734

Relnote: "Adding Material3 Stepper to Wear Compose. This version doesn't
	have range semantics by default. It can be added using modifier
	such as Modifier.rangeSemantics."

Change-Id: Ic39fddeb8f959bdff6aee0ee6b2cbe6c906feab8