Integrate setNextScheduleTimeOverride with TestDriver
Uses setPeriodMet() to unblock any override schedule, even on the first
run. First runs of periodic work with a schedule override don't use
their initialDelay in production workmanager, so the
setinitialdelayMet() method does nothing to them in test.

When work is ready to run, TestScheduler now has the ability to roll
back the nextschedule to allow it to run.

Test: Added new tests.
Relnote: "Added ability to bypass overrideNextScheduleTime using
TestDriver and fixed issues with testability"

Change-Id: Ic29056d338f5aeb6c1825f487d8f206908465141