Skip recomposition of subcompositions that will be removed
Compose does not ensure repeatable read consistency across a sub-composition boundary (that is we do not guarantee that a mutable value read by the parent composition will be the same value read by the sub-composition). However in cases where we can detect that a subcomposition will be removed in the next applyChanges, we can skip recomposing that
subcomposition.

This is accomplished by keeping a map of removed compositions reported via the composition context.
When a composition is marked for deletion, it checks through its marked groups for any nested composition contexts. For each one it finds, reports its deletion to the recomposer via the context. When the nested composition is
recomposed, if the recomposers map of removed compositions contains the composition, it is skipped.

This change is an improvement on current behaviour but could be improved by ensuring parent compositions are always recomposed before any nested compositions. This mostly already occurs but making this enforced would further improve this situation and is why the linked bug is not yet considered fixed.

Bug: 254645321
Relnote: Skip recomposition of subcompositions that will be removed
Test: testSubcompositionDisposedInParent
Change-Id: Ieeb9919897a9f9b4274ddc77e66608a673cd1112