Fix the center top app bar alignment
Fix the alignement of the center top app bar to ensure its title does
not render over the icons when it's very long and the width of the
action icons is different than the width of the navigation icon.

This fix gives priority to the title's text display and shifts the title
from its center position as needed to show as much text as possible.

Fixes: 236994621
Test: AppBarTest & manual on LTR and RTL layouts
Relnote: "Fix the alignement of the center top app bar to ensure that a
long title doesn't render over the action icons or the navigation icon."

Change-Id: I4369f6c51c711129bb30b1de58684131b02499ae